Renamed the test files inside boards/imix/src to be more descriptive.
