+++ boards/imix/src/main.rs
@@ -39,11 +39,13 @@ pub mod io;
 #[allow(dead_code)]
 mod i2c_dummy;
 #[allow(dead_code)]
-mod lowpan_frag_dummy;
-#[allow(dead_code)]
 mod spi_dummy;
 #[allow(dead_code)]
-mod app_layer_icmp_lowpan_frag;
+mod ipv6_lowpan_test;
+#[allow(dead_code)]
+mod udp_lowpan_test;
+#[allow(dead_code)]
+mod icmp_lowpan_test;
 
 #[allow(dead_code)]
 mod aes_test;
